+++ apps/src/templates/teacherDashboard/RosterDialog.jsx
@@ -7,10 +7,10 @@ import {PLATFORMS} from '@cdo/apps/metrics/AnalyticsConstants.js';
 import analyticsReporter from '@cdo/apps/metrics/AnalyticsReporter';
 import locale from '@cdo/locale';
 
-import RailsAuthenticityToken from '../../lib/util/RailsAuthenticityToken';
 import color from '../../util/color';
 import BaseDialog from '../BaseDialog';
 
+import ReauthorizeGoogleClassroom from './ReauthorizeGoogleClassroom';
 import {classroomShape, loadErrorShape} from './shapes';
 import {
   cancelImportRosterFlow,
@@ -143,19 +143,6 @@ LoadError.propTypes = {
   rosterProvider: PropTypes.string,
   loginType: PropTypes.string,
 };
-
-const REAUTHORIZE_URL =
-  '/users/auth/google_oauth2?scope=userinfo.email,userinfo.profile,classroom.courses.readonly,classroom.rosters.readonly';
-function ReauthorizeGoogleClassroom() {
-  return (
-    <form method="POST" action={REAUTHORIZE_URL}>
-      <RailsAuthenticityToken />
-      <button type="submit" style={ctaButtonStyle}>
-        {locale.authorizeGoogleClassrooms()}
-      </button>
-    </form>
-  );
-}
 
 class RosterDialog extends React.Component {
   static propTypes = {
